refactor: changing archives paths
