+++ lib/details/widgets/asset_graph.dart
@@ -1,13 +1,13 @@
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import '../controller/details_asset_provider.dart';
-import '../controller/x_axis_provider.dart';
+
 import '../../shared/models/asset_model.dart';
 import '../../shared/templates/app_assets.dart';
 import '../../shared/utils/currency_formater.dart';
-
 import '../../shared/utils/decimal_to_double.dart';
+import '../controller/details_asset_notifier_provider.dart';
+import '../controller/x_axis_provider.dart';
 
 class GraphDetails extends StatefulHookConsumerWidget {
   const GraphDetails({Key? key}) : super(key: key);
