+++ lib/details/widgets/change_x_axis_list.dart
