+++ lib/details/widgets/info_card_asset.dart
