+++ lib/pages/view/pages.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:projeto_criptos/details/view/details_screen.dart';
 
 import '../widgets/cripto_bottom_navigation_bar.dart';
-import '../widgets/page_index_provider.dart';
+import '../controller/page_index_provider.dart';
 import '../widgets/page_view_controller.dart';
 
 class Pages extends StatefulHookConsumerWidget {
   const Pages({Key? key}) : super(key: key);
-  static String route = '/';
   @override
   ConsumerState<Pages> createState() => _PagesState();
 }
@@ -33,7 +33,10 @@ class _PagesState extends ConsumerState<Pages> {
     index = ref.read(pageIndexProvider.state).state;
     controller = PageController(initialPage: index);
   }
-
+  List<Widget> pages = [
+    const Pages(),
+    const DetailsScreen(),
+  ];
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -42,7 +45,6 @@ class _PagesState extends ConsumerState<Pages> {
         onPageChanged: changePage,
       ),
       bottomNavigationBar: CriptoBottomNavigationBar(
-        controller: controller,
         onTap: changePage,
       ),
     );
