+++ lib/shared/routes/app_routes.dart
@@ -4,6 +4,6 @@ import '../../details/view/details_screen.dart';
 import '../../pages/view/pages.dart';
 
 Map<String, Widget Function(BuildContext)> appRoutes = {
-  DetailsScreen.route: (context) => const DetailsScreen(),
-  Pages.route: (context) => const Pages(),
+  '/details': (context) => const DetailsScreen(),
+  '/': (context) => const Pages(),
 };
